==> PD1_Treat/01b_tenX/01a_Preprocessing_Immune.Rmd <==
---
title: "R Notebook"
output: html_document
---

## Preprocess matrices to seurat object, then do broad annotation (myeloid, tcell) + detailed (cd4, cd8) and save these seurat objects

```{r}
knitr::opts_chunk$set(echo = TRUE)

base_dir = "/Users/jlabelle/"
working_dir<- paste0(base_dir, 
                     "Dropbox (Partners HealthCare)/FilbinLab/",
                     "data_analysis/Projects/git_repos/Pedi_Immune/PD1_Treat/01b_tenX/")
script_dir = "Dropbox (Partners HealthCare)/FilbinLab/data_analysis/Scripts/"


preprocessing = paste0(base_dir, script_dir, "single_cell_preprocessing_helper_functions.R")
source(preprocessing)
library(Matrix)

analysis_dir<- "analysis/preprocessing/"
figure_dir<- "figures/preprocessing/"
data_dir<- "data/"

if(!dir.exists(analysis_dir)){dir.create(analysis_dir, recursive = TRUE)}
if(!dir.exists(figure_dir)){dir.create(figure_dir, recursive = TRUE)}

marker_dir<- paste0(base_dir, "Dropbox (Partners HealthCare)/FilbinLab/data_analysis/Shared/Marker_genes/")
hkgenes <- read.table(paste0(marker_dir, "tirosh_house_keeping.txt"), skip = 2)
hkgenes <- as.vector(hkgenes$V1)
```

## Read in immune matrixes- 1 per sample
```{r}
samples<- list.files(data_dir)

all_matrix<- lapply(samples, function(x){
  print(x)
  matrix_dir<- paste0(data_dir, x, "/Immune")
  
  ## tumor/immune as available
  if(dir.exists(matrix_dir)){
    barcode.path <- paste0(matrix_dir, "/filtered_feature_bc_matrix/barcodes.tsv.gz")
    features.path <- paste0(matrix_dir, "/filtered_feature_bc_matrix/features.tsv.gz")
    matrix.path <- paste0(matrix_dir, "/filtered_feature_bc_matrix/matrix.mtx.gz")
    mat <- readMM(file = matrix.path)
    feature.names = read.delim(features.path,
                               header = FALSE,
                               stringsAsFactors = FALSE)
    barcode.names = read.delim(barcode.path,
                               header = FALSE,
                               stringsAsFactors = FALSE)
    colnames(mat) = barcode.names$V1
    rownames(mat) = feature.names$V2
  } else{next}
  return(mat)
})
names(all_matrix)<- samples

## Some cell names are duplicated. add sample name to clarify
all_matrix<- lapply(names(all_matrix), function(x){
  mtx<- all_matrix[[x]]
  colnames(mtx)<- paste0(x, "_", colnames(mtx))
  return(mtx)
})
names(all_matrix)<-samples

## Replace any "-" or "." with "_"- will do same for TCR. To ensure cell names match.
all_matrix<- lapply(names(all_matrix), function(x){
  mtx<- all_matrix[[x]]
  colnames(mtx)<-gsub("-", "_", gsub("\\.", "_", colnames(mtx)))
  return(mtx)
})
names(all_matrix)<-samples
```

## Make simple metadata for each matrix- cells + sample
```{r}
all_meta<- lapply(names(all_matrix), function(x){
  matrix<- all_matrix[[x]]
  meta<- data.frame(cell=colnames(matrix), sample=rep(x, ncol(matrix)))
  return(meta)
})
```

## Save cm and meta
```{r}
meta<- do.call("rbind", all_meta)

lapply(all_matrix, function(x){sum(!rownames(x) == rownames(all_matrix[[1]]))})
cm<- do.call("cbind", all_matrix)

## There are 95 genes that are duplicated (same for both samples)-sum counts
## I'm pretty sure seurat will do this automatically for you- but it throws a warning and I don't like that
  ## Get cm of genes with dup genes
  dup_genes<- rownames(cm)[duplicated(rownames(cm))]
  dup_cm<-cm[rownames(cm) %in% dup_genes,]
  dup_genes<- rownames(dup_cm)
  dup_cm<- as.data.frame(dup_cm)
  dup_cm$gene<- dup_genes

  ## Group by gene and get total counts across cell
  dup_cm<- dup_cm %>% group_by(gene) %>% summarise(across(everything(), list(sum))) %>% as.data.frame()
  rownames(dup_cm)<-dup_cm$gene
  dup_cm$gene<- NULL
  colnames(dup_cm)<- gsub("_1$", "", colnames(dup_cm)) ## removing _1 added on in grouping process
  
  ## Add back to full cm
  cm_unDup<- as.data.frame(cm[!rownames(cm) %in% rownames(dup_cm),])
  cm<- rbind(cm_unDup, dup_cm)

saveRDS(cm, file=paste0(analysis_dir, "raw_cm.Rds"))
saveRDS(meta, file=paste0(analysis_dir, "raw_meta.Rds"))
```

## Create seurat object, do initial QC
```{r}
cm<- readRDS(paste0(analysis_dir, "raw_cm.Rds"))
meta<- readRDS(paste0(analysis_dir, "raw_meta.Rds"))

seurat_raw<- RunFullSeurat(cm=cm, RunHarmony = FALSE, samples=meta$sample)


## Identify # of HK genes
hkgenes.notfound.indices <- which(!hkgenes %in% rownames(seurat_raw))  
if (length(hkgenes.notfound.indices)) {
  message = paste0(paste0(hkgenes[hkgenes.notfound.indices], collapse = " "), " housekeeping genes were not found")
  print(message)
}

# Get number of expressed housekeeping genes.
hkgenes.found <- hkgenes[-hkgenes.notfound.indices]  # remove hkgenes that were not found
seurat_raw[["nHK_RNA"]] <- Matrix::colSums(seurat_raw[['RNA']]@data[hkgenes.found, ] > 0)


## Get percentage of mito genes
seurat_raw[["percent_mito"]] <- PercentageFeatureSet(seurat_raw, pattern = "^MT-")
  


## QC plots
p1<-DimPlot(seurat_raw, group.by="sample", label=TRUE)+NoLegend()
p2<- VlnPlot(seurat_raw, features = "nFeature_RNA", group.by="sample", pt.size = 0)+NoLegend()
p3<- VlnPlot(seurat_raw, features="percent_mito", group.by="sample", pt.size = 0)+NoLegend()
p4<- VlnPlot(seurat_raw, features="nHK_RNA", group.by="sample", pt.size = 0)+NoLegend()


p1+p2+p3+p4
ggsave(paste0(figure_dir, "nFeature.Count_BySample.png"), width=8, height=10)
saveRDS(seurat_raw, file=paste0(analysis_dir, "raw_seurat.Rds"))
```

## Plot histogram of nFeature/nHK/nMitochondrial to determine good cutoff
```{r}
seurat_raw<- readRDS(paste0(analysis_dir, "raw_seurat.Rds"))

feature_thresh<- 500
hk_thresh<- 50

df<- seurat_raw@meta.data
p1<- ggplot(df, aes(x=nFeature_RNA, fill=sample))+geom_histogram()+ggtitle("nFeatures")+geom_vline(xintercept = feature_thresh)
p2<- ggplot(df, aes(x=nHK_RNA, fill=sample))+ geom_histogram()+ggtitle("n HK genes")+theme(legend.position = "none")+geom_vline(xintercept = hk_thresh)
p3<- ggplot(df, aes(x=nCount_RNA, fill=sample))+ geom_histogram() + ggtitle("nCounts")
p4<- ggplot(df, aes(x=percent_mito, fill=sample))+ geom_histogram() + ggtitle("percent mito")+theme(legend.position = "none")
p1+p2+p3+p4
ggsave(paste0(figure_dir, "Histogram_Features.Counts.HK.Mito.png"), width=10, height=8)
```




## Filter cells by nFeature + nHK
```{r}
seurat_raw<- readRDS(paste0(analysis_dir, "raw_seurat.Rds"))

## Filter 
seurat_obj_filt <- subset(seurat_raw, subset = nFeature_RNA > feature_thresh  & nHK_RNA > hk_thresh)
seurat_raw$PF<- ifelse(colnames(seurat_raw) %in% colnames(seurat_obj_filt), "Pass", "Fail")

df<- seurat_raw@meta.data
df<- as.data.frame(table(df$sample, df$PF))
colnames(df)<- c("Sample", "PF", "nCells")
ggplot(df, aes(x=Sample, y=nCells, fill=PF))+
  geom_bar(stat="identity", position="stack")+
  theme(axis.text.x = element_text(angle=45, hjust=1))+
  ggtitle(paste0("nFeature_RNA > ",feature_thresh, 
                 "\nnHK_RNA > ", hk_thresh))+
  xlab("")
ggsave(paste0(figure_dir, "PF_BySample_nFeature.nHK.png"), width=3, height=3)

## recluster without low quality cells
seurat_obj_filt_meta<- seurat_obj_filt@meta.data
seurat_obj_filt_meta$batch<- seurat_obj_filt_meta$sample
seurat_obj_filt_meta$sample<- gsub("_new", "", seurat_obj_filt_meta$sample)
seurat_obj_filt<- RunFullSeurat(seurat_obj_filt@assays$RNA@counts, RunHarmony = TRUE, samples=seurat_obj_filt_meta$sample)
seurat_obj_filt<-AddMetaData(seurat_obj_filt, seurat_obj_filt_meta[,!colnames(seurat_obj_filt_meta) %in% colnames(seurat_obj_filt@meta.data)])

## Make sure no low qual clusters
FeaturePlot(seurat_obj_filt, features=c("nFeature_RNA", "nHK_RNA", "percent_mito"), cols=c("grey", "red"))
ggsave(paste0(figure_dir, "UMAP_QCmetrics_postFiltration.png"), width=6, height=5.5)

saveRDS(seurat_obj_filt, file=paste0(analysis_dir, "seurat_obj.Rds"))
saveRDS(seurat_raw, file=paste0(analysis_dir, "raw_seurat.Rds"))
```

## Identify any non-immune normal cells (AC, OC, tumor)
```{r}
seurat_obj<- readRDS(paste0(analysis_dir, "seurat_obj.Rds"))

## Reduce resolution to identify markers for major clusters
seurat_obj<-FindClusters(seurat_obj, resolution = 0.1)
DimPlot(seurat_obj, label=TRUE)
seurat_obj<- SetIdent(seurat_obj, value=seurat_obj$seurat_clusters)
DimPlot(seurat_obj, group.by = "seurat_clusters", label=TRUE)+NoLegend()+
  FeaturePlot(seurat_obj, features=c("CD3G", "CSF1R", "CD8A", "FOXP3", "SPARCL1", "SOX3"))
ggsave(paste0(figure_dir, "UMAP_seurat.clusters_majorClusters.png"), width=12, height=6)

## Find markers for major clusters
all_markers<- FindAllMarkers(seurat_obj, min.pct = 0.5, logfc.threshold = 0.5)
all_markers_filtered<- all_markers[all_markers$p_val_adj<0.05,]
top_markers<- all_markers_filtered %>% group_by(cluster) %>% top_n(n=10, wt=avg_log2FC) %>% as.data.frame()
write.csv(all_markers_filtered, file=paste0(analysis_dir, "major_clusters_markers_filtered.csv"))


## Assign cluster DC/tumor clusters
seurat_obj$immune.other_annot<- ifelse(seurat_obj$seurat_clusters==6, "DC", ifelse(seurat_obj$seurat_clusters==4, "Tumor", "Immune"))
FeaturePlot(seurat_obj, features=c("CD3G", "CSF1R", "SPARCL1", "SOX3", "FCN1")) + DimPlot(seurat_obj, group.by = "immune.other_annot")
ggsave(paste0(figure_dir, "UMAP_removeAC.Tumor.png"), width=7, height=6)


## Remove DC/Tumor
seurat_obj<- subset(seurat_obj, immune.other_annot=="Immune")
meta<- seurat_obj@meta.data
seurat_obj<- RunFullSeurat(seurat_obj@assays$RNA@counts, RunHarmony = FALSE, samples=meta$sample)
seurat_obj<- AddMetaData(seurat_obj, meta[,!colnames(meta) %in% colnames(seurat_obj@meta.data)])

saveRDS(seurat_obj, file="analysis/preprocessing/immune_seurat.Rds")
```

## Convert scale to TPM- to match with SS2 
May not make a difference for projection, but likely best to have identical scale for raw counts when comparing SS2/tenX
```{r}
seurat_obj<- readRDS("analysis/preprocessing/immune_seurat.Rds")
seurat_meta<- seurat_obj@meta.data

## Convert to CPM
cm<- seurat_obj@assays$RNA@counts
cm1<- apply(cm, 2, function(x){x/sum(x) * 1e06}) ## run on O2

## Rerun clustering (just so processing is completely identical to ss2)
seurat_obj<- RunFullSeurat(cm, RunHarmony = TRUE, samples=seurat_obj$sample)
seurat_obj<- AddMetaData(seurat_obj, seurat_meta[,!colnames(seurat_meta) %in% colnames(seurat_obj@meta.data)])

saveRDS(seurat_obj, file="analysis/preprocessing/immune_seurat.Rds")
```
