==> PD1_Treat/01b_tenX/02_ProjectBroad.DetailedAnnots.Rmd <==
---
title: "ProjectBroadAnnots.Rmd"
author: "Jenna LaBelle"
date: "10/07/22"
output: html_document
---

## Project "broad" (T cell and myeloid) and "detailed" (CD4, CD8, DC, Monocyte, Bcell) annotations from the full immune cohort to PD1 treated patients

```{r}
library(Seurat) 

base_dir = "/Users/jlabelle/"
script_dir = "Dropbox (Partners HealthCare)/FilbinLab/data_analysis/Shared/Scripts/"
working_dir<- paste0(base_dir, 
                     "Dropbox (Partners HealthCare)/FilbinLab/",
                     "data_analysis/Projects/git_repos/Pedi_Immune/PD1_Treat/01b_tenX/")

analysis_dir<- paste0(working_dir, "/analysis/project_broad.detailed_annot/")
figure_dir<- paste0(working_dir, "/figures/project_broad.detailed_annot/")
full_cohort_dir<- "../../ImmuneCellAnalysis/"

if(!dir.exists(figure_dir)){dir.create(figure_dir, recursive = TRUE)}
if(!dir.exists(analysis_dir)){dir.create(analysis_dir, recursive = TRUE)}

preprocessing = paste0(base_dir, script_dir, "single_cell_preprocessing_helper_functions.R")
nmf_helper = paste0(base_dir, script_dir, "NMF_helper_function.R")
source(preprocessing)
source(nmf_helper)
source(paste0(base_dir, script_dir, "seurat_transferAnchors_helper.R"))
```



## Load seurat objects: pd-1 (processed in 01b_Preprocess_QC_ToSeurat.Rmd) and full immune cohort (both myeloid and t cells)
```{r}

pd1_seurat<- readRDS( "analysis/preprocessing/immune_seurat.Rds")
full_seurat<- readRDS(paste0(full_cohort_dir,
                             "01_Preprocessing/02_Counts_Preprocess/analysis/detailed_annotation/pedOnly_nomuv63/",
                             "tcell.myeloid_seurat_harmony.Rds"))
full_tcell_seurat<- readRDS(paste0(full_cohort_dir,
                                   "02a_Tcells/02_Ped.Adult/03_Visualize.Reannotate_programs/analysis/Reannotate_programs/",
                                   "rank8_sharedByunion/",
                                   "ped_seurat_tcell_harmony.Rds"))
full_myeloid_seurat<- readRDS(paste0(full_cohort_dir,
                                     "01_Preprocessing/02_Counts_Preprocess/analysis/detailed_annotation/",
                                     "pedOnly_nomuv63/",
                                     "myeloid_seurat_harmony.Rds"))

## Colors 
immune_colors<- readRDS(paste0(full_cohort_dir, "/plot_colors/ImmuneCell.celltypes.Rds"))

```


## Projection of myeloid/tcell annotations with pcproject 
```{r}
pd1_seurat<- my_project_seurat(full_seurat, pd1_seurat, reference_annotation = "broad_annot",
                               prediction_colNames = c("broad_annot_projected", "Tcell_score_projected", 
                                                             "Myeloid_score_projected", "max_score_projected"))


## plot
DimPlot(pd1_seurat, group.by="broad_annot_projected", pt.size = 1, label=TRUE, label.size = 6)+ NoLegend()+
  scale_color_manual(values=immune_colors[names(immune_colors) %in% pd1_seurat$broad_annot_projected])
ggsave(paste0(figure_dir, "UMAP_broadAnnot_projectedFromFull.png"), width=6, height=6)

## Confirm high expression of myeloid/tcell markers by predicted broad annotation
cowplot::plot_grid(plotlist=lapply(c("CSF1R", "CD3G", "CD8A", "CD4"), function(x){
  VlnPlot(pd1_seurat, features=x, group.by = "broad_annot_projected", pt.size=0)+ NoLegend()+
    scale_fill_manual(values=immune_colors[names(immune_colors) %in% pd1_seurat$broad_annot_projected])+
    xlab("")
}), ncol=4)
ggsave(paste0(figure_dir, "Vln_broadMarkers_projectedFromFull.png"), width=12, height=4)

## Proportion
my_barchart_programProp(pd1_seurat, coloring_variable = "broad_annot_projected", colors=immune_colors)
ggsave(paste0(figure_dir, "Barchart_broadAnnot_bySample.png"), width=4, height=4)

saveRDS(pd1_seurat, file="analysis/preprocessing/immune_seurat.Rds")
```

## Split into myeloid/tcell seurat objects, rerun processing
```{r}
pd1_seurat<- readRDS("analysis/preprocessing/immune_seurat.Rds")

## Myeloid
myeloid_pd1<- subset(pd1_seurat, broad_annot_projected=="Myeloid")
myeloid_pd1_meta<- myeloid_pd1@meta.data
myeloid_pd1<- RunFullSeurat(myeloid_pd1@assays$RNA@counts, samples= myeloid_pd1$sample, RunHarmony = TRUE)
myeloid_pd1<- AddMetaData(myeloid_pd1, myeloid_pd1_meta[,!colnames(myeloid_pd1_meta) %in% colnames(myeloid_pd1@meta.data)])

## Tcell
tcell_pd1<- subset(pd1_seurat, broad_annot_projected=="Tcell")
tcell_pd1_meta<- tcell_pd1@meta.data
tcell_pd1<- RunFullSeurat(tcell_pd1@assays$RNA@counts, samples= tcell_pd1$sample, RunHarmony = TRUE)
tcell_pd1<- AddMetaData(tcell_pd1, tcell_pd1_meta[,!colnames(tcell_pd1_meta) %in% colnames(tcell_pd1@meta.data)])

## Check for expression of expected markers
  ## Myeloid: CD3+ macrophages present
  FeaturePlot(myeloid_pd1, features=c("CSF1R", "CD3G")) 
  ggsave(paste0(figure_dir, "Myeloid_broad_markers_preReannotation.png"), width=10, height=5)
  
  ## Tcell: CD3+ macrophages also present- remove and add into myeloid
  FeaturePlot(tcell_pd1, features=c("CSF1R", "CD3G")) 
  ggsave(paste0(figure_dir, "Tcell_broad_markers_preReannotation.png"), width=10, height=5)
  
  ## Remove CD3+ macrophages from T cell and add to myeloid
  tcell_cd3.myeloid<- subset(tcell_pd1, seurat_clusters==10)
  myeloid_pd1<- merge(tcell_cd3.myeloid, myeloid_pd1)
  tcell_pd1<- subset(tcell_pd1, seurat_clusters != 10)
  
  ## Rerun clustering
  myeloid_pd1_meta<- myeloid_pd1@meta.data
  myeloid_pd1<- RunFullSeurat(myeloid_pd1@assays$RNA@counts, samples= myeloid_pd1$sample, RunHarmony = TRUE)
  myeloid_pd1<- AddMetaData(myeloid_pd1, myeloid_pd1_meta[,!colnames(myeloid_pd1_meta) %in% colnames(myeloid_pd1@meta.data)])
  myeloid_pd1$broad_annot_projected_manual<- "Myeloid"
  
  tcell_pd1_meta<- tcell_pd1@meta.data
  tcell_pd1<- RunFullSeurat(tcell_pd1@assays$RNA@counts, samples= tcell_pd1$sample, RunHarmony = TRUE)
  tcell_pd1<- AddMetaData(tcell_pd1, tcell_pd1_meta[,!colnames(tcell_pd1_meta) %in% colnames(tcell_pd1@meta.data)])
  
  ## Confirm expression of expected markers
  FeaturePlot(myeloid_pd1, features=c("CSF1R", "CD3G")) 
  ggsave(paste0(figure_dir, "Myeloid_broad_markers.png"), width=10, height=5)
  
  FeaturePlot(tcell_pd1, features=c("CSF1R", "CD3G")) 
  ggsave(paste0(figure_dir, "Tcell_broad_markers.png"), width=10, height=5)

saveRDS(myeloid_pd1, file=paste0(analysis_dir, "myeloid_seurat.Rds"))
saveRDS(tcell_pd1, file=paste0(analysis_dir, "tcell_seurat.Rds"))
```

## Projection of DC/BC/Monocyte annotations for myeloid cells with pcproject 
```{r}
myeloid_pd1<- readRDS(paste0(analysis_dir, "myeloid_seurat.Rds"))

myeloid_pd1<- my_project_seurat(full_myeloid_seurat, myeloid_pd1, reference_annotation = "detailed_annot",
                                prediction_colNames =c("detailed_annot_projected", "Myeloid_score_projected", 
                                                             "Bcell_score_projected", "DC_score_projected",
                                                             "max_Detailed_score_projected") )


## plot
DimPlot(myeloid_pd1, group.by="detailed_annot_projected", pt.size = 1, label=TRUE, label.size = 6)+ NoLegend()+
  scale_color_manual(values=immune_colors[names(immune_colors) %in% myeloid_pd1$detailed_annot_projected])
ggsave(paste0(figure_dir, "UMAP_myeloid_detailedAnnot_projectedFromFull.png"), width=6, height=6)

## Very few Bcells and no dendritic cells predicted based on projection. Seem to not be B cells?
FeaturePlot(myeloid_pd1, features=c("CD207", "CD19", "IL3RA", "CLEC4C", "CSF1R", "CD3G"))
ggsave(paste0(figure_dir, "UMAP_Myeloid_detailedMarkers.png"), width=6, height=8)
```




## Projection of CD4/CD8 annotations for T cells with pcproject 
```{r}
tcell_pd1<- readRDS(paste0(analysis_dir, "tcell_seurat.Rds"))

tcell_pd1<- my_project_seurat(full_tcell_seurat, tcell_pd1, reference_annotation = "detailed_annot_byNMFrank8_mergedByunion",
                                prediction_colNames =c("detailed_annot_projected", "CD4_score_projected", 
                                                             "CD8_score_projected", "max_Detailed_score_projected"))


## plot
DimPlot(tcell_pd1, group.by="detailed_annot_projected", pt.size = 1, label=TRUE, label.size = 6)+ NoLegend()+
  scale_color_manual(values=immune_colors[names(immune_colors) %in% tcell_pd1$detailed_annot_projected])
ggsave(paste0(figure_dir, "UMAP_Tcell_detailedAnnot_projectedFromFull.png"), width=6, height=6)

## Confirm high expression of tcell markers by predicted detailed annotation
cowplot::plot_grid(plotlist=lapply(c( "CD3G", "CD8A", "CD4"), function(x){
  VlnPlot(tcell_pd1, features=x, group.by = "detailed_annot_projected")+ NoLegend()+
    scale_fill_manual(values=immune_colors[names(immune_colors) %in% tcell_pd1$detailed_annot_projected])+
    xlab("")
}), ncol=3)
ggsave(paste0(figure_dir, "Vln_Tcell_detailedMarkers_projectedFromFull.png"), width=9, height=4)

FeaturePlot(tcell_pd1, features=c("CD3G", "CD4", "CD8A"), cols = c("grey", "red"), ncol=3)
ggsave(paste0(figure_dir, "UMAP_Tcell_detailedMarkers.png"), width=12, height=3)

## Save with annotations
saveRDS(tcell_pd1, file=paste0(analysis_dir, "tcell_seurat.Rds"))

## Split into cd4/cd8, recluster
cd4_pd1<- subset(tcell_pd1, detailed_annot_projected=="CD4")
cd4_pd1_meta<- cd4_pd1@meta.data
cd4_pd1<- RunFullSeurat(cd4_pd1@assays$RNA@counts, samples=cd4_pd1$sample, RunHarmony = TRUE)
cd4_pd1<- AddMetaData(cd4_pd1, cd4_pd1_meta[,!colnames(cd4_pd1_meta) %in% colnames(cd4_pd1@meta.data)])

cd8_pd1<- subset(tcell_pd1, detailed_annot_projected=="CD8")
cd8_pd1_meta<- cd8_pd1@meta.data
cd8_pd1<- RunFullSeurat(cd8_pd1@assays$RNA@counts, samples=cd8_pd1$sample, RunHarmony = TRUE)
cd8_pd1<- AddMetaData(cd8_pd1, cd8_pd1_meta[,!colnames(cd8_pd1_meta) %in% colnames(cd8_pd1@meta.data)])

## Both CD4 and CD8 show misannotated cells- fix annotation and recluster
p1<-FeaturePlot(cd8_pd1, features=c("CD4", "CD8A"), cols=c("grey", "red"))
p2<-FeaturePlot(cd4_pd1, features=c("CD4", "CD8A"), cols=c("grey", "red"))
cowplot::plot_grid(plotlist=list(p1, p2), ncol=1)
ggsave(paste0(figure_dir,"UMAP_CD4.CD8_detailedMarkers_preReannotation.png"), width=7, height=6)

  ## Fix CD4: 2 clear CD8 cluster + 1 with CD4/CD8 to subcluster + 1 cycling cluster annotated cell wise
  DimPlot(cd4_pd1, label=TRUE) + VlnPlot(cd4_pd1, features=c("CD4", "CD8A"))
  
    ## clear clusters
    cd4_pd1$detailed_annot_projected_manual<- ifelse(cd4_pd1$seurat_clusters %in% c(4,6), "CD8", cd4_pd1$detailed_annot_projected) 
    
    ## subclustering
    subcluster<- subset(cd4_pd1, seurat_clusters ==5)
    subcluster<- RunFullSeurat(subcluster@assays$RNA@counts, RunHarmony = TRUE, samples=subcluster$sample,
                               pca_dims = 95, dims = 95)
    subcluster$detailed_annot_projected_manual<- gsub(1, "CD8", gsub(0, "CD4", subcluster$seurat_clusters))
    cd4_pd1$detailed_annot_projected_manual<- ifelse(colnames(cd4_pd1) %in% colnames(subcluster),
                                                     as.character(plyr::mapvalues(colnames(cd4_pd1),
                                                                                  colnames(subcluster),
                                                                                  subcluster$detailed_annot_projected_manual,
                                                                                  warn_missing = FALSE)),
                                                     cd4_pd1$detailed_annot_projected_manual)
    
    ## cycling annoated cell wise
    cd4_pd1$cd8_expr<- as.data.frame(cd4_pd1@assays$RNA@counts["CD8A",])
    cd4_pd1$detailed_annot_projected_manual<- ifelse(cd4_pd1$seurat_clusters==7 & cd4_pd1$cd8_expr > 0, "CD8",
                                                     cd4_pd1$detailed_annot_projected_manual)
  
  DimPlot(cd4_pd1, label=TRUE) + 
    FeaturePlot(cd4_pd1, features=c("CD4", "CD8A"), cols=c("grey", "red"))+
    VlnPlot(cd4_pd1, features=c("CD4", "CD8A"))+
    DimPlot(cd4_pd1, group.by = "detailed_annot_projected_manual")
  ggsave(paste0(figure_dir, "CD4_reannotateToCD8.png") ,width = 10, height=10)
  
  
  ## Fix CD8: 1 clear CD4 cluster + 1 cycling cluster annotated cell wise
  DimPlot(cd8_pd1, label=TRUE) + VlnPlot(cd8_pd1, features=c("CD4", "CD8A"))
  
    ## clear cluster
    cd8_pd1$detailed_annot_projected_manual<- ifelse(cd8_pd1$seurat_clusters==1, "CD4", cd8_pd1$detailed_annot_projected)
    
    ## cell wise for cycling
    cd8_pd1$cd4_expr<- as.data.frame(cd8_pd1@assays$RNA@counts["CD4",])
    cd8_pd1$detailed_annot_projected_manual<- ifelse(cd8_pd1$seurat_clusters%in%c(8) & cd8_pd1$cd4_expr > 0, "CD4",
                                                     cd8_pd1$detailed_annot_projected_manual)
  
  DimPlot(cd8_pd1, label=TRUE) + 
    FeaturePlot(cd8_pd1, features=c("CD4", "CD8A"), cols=c("grey", "red"))+
    VlnPlot(cd8_pd1, features=c("CD4", "CD8A"))+
    DimPlot(cd8_pd1, group.by = "detailed_annot_projected_manual")
  ggsave(paste0(figure_dir, "CD8_reannotateToCD4.png") ,width = 10, height=10)
  
  ## Add CD4/CD8 to correct object and recluster
  cd8_pd1_reAnnot<- my_mergeRerunClustering(subset(cd8_pd1, detailed_annot_projected_manual=="CD8"),
                                            subset(cd4_pd1, detailed_annot_projected_manual=="CD8"), 
                                            Integration="Harmony", integrateBy = "sample")
  cd4_pd1_reAnnot<- my_mergeRerunClustering(subset(cd8_pd1, detailed_annot_projected_manual=="CD4"),
                                            subset(cd4_pd1, detailed_annot_projected_manual=="CD4"), 
                                            Integration="Harmony", integrateBy = "sample")
  
  ## Confirm finalized annots
  p1<-FeaturePlot(cd8_pd1_reAnnot, features=c("CD4", "CD8A"), cols=c("grey", "red"))
  p2<- FeaturePlot(cd4_pd1_reAnnot, features=c("CD4", "CD8A"), cols=c("grey", "red"))
  cowplot::plot_grid(plotlist=list(p1, p2), ncol=1)
  ggsave(paste0(figure_dir, "UMAP_Tcells_detailedMarkers_postReannotation.png"), width=6, height=6)
  
  ## Add back to full tcell
  tcell_pd1$detailed_annot_projected_manual<- ifelse(colnames(tcell_pd1) %in% colnames(cd4_pd1_reAnnot), "CD4",
                                                     ifelse(colnames(tcell_pd1) %in% colnames(cd8_pd1_reAnnot), "CD8", "issue"))
  DimPlot(tcell_pd1, group.by = "detailed_annot_projected_manual")+
    scale_color_manual(values = immune_colors[names(immune_colors) %in% unique(tcell_pd1$detailed_annot_projected_manual)])
  ggsave(paste0(figure_dir, "UMAP_Tcells_detailedAnnot_postReannotation.png"), width=6, height=5)
  

saveRDS(cd4_pd1_reAnnot, file=paste0(analysis_dir, "cd4_seurat.Rds"))
saveRDS(cd8_pd1_reAnnot, file=paste0(analysis_dir, "cd8_seurat.Rds"))
saveRDS(tcell_pd1, file=paste0(analysis_dir, "tcell_seurat.Rds"))
```
