==> PD1_Treat/01b_tenX/05_ProgramDE.Rmd <==
---
title: "DE between PD1/Untreated programs"
author: "Jenna LaBelle"
date: "10/07/22"
output: html_document
---

DE analysis comparing each program in PD1 vs Untreated

```{r}
library(Seurat) 
library(ggalluvial)

base_dir = "/Users/jlabelle/"
script_dir = "Dropbox (Partners HealthCare)/FilbinLab/data_analysis/Shared/Scripts/"
working_dir<- paste0(base_dir, 
                     "Dropbox (Partners HealthCare)/FilbinLab/",
                     "data_analysis/Projects/git_repos/Pedi_Immune/PD1_Treat/01b_TenX/")

analysis_dir<- paste0(working_dir, "/analysis/program_de/")
figure_dir<- paste0(working_dir, "/figures/program_de/")
data_dir<- paste0(working_dir, "analysis/Merge/")
full_cohort_dir<- "../../ImmuneCellAnalysis/"

if(!dir.exists(figure_dir)){dir.create(figure_dir, recursive = TRUE)}
if(!dir.exists(analysis_dir)){dir.create(analysis_dir, recursive = TRUE)}

preprocessing = paste0(base_dir, script_dir, "single_cell_preprocessing_helper_functions.R")
nmf_helper = paste0(base_dir, script_dir, "NMF_helper_function.R")
source(preprocessing)
source(nmf_helper)
source(paste0(base_dir, script_dir, "seurat_transferAnchors_helper.R"))

```


## Load seurat objects: pd-1 (processed in 01b_Preprocess2QC_ToSeurat.Rmd) and full immune cohort (both myeloid and t cells)

```{r}
## Load in tenx/ss2 merged seurat objects
files<- list.files(data_dir, pattern="seurat")
seurat_list<- lapply(files, function(f){
   readRDS(paste0(data_dir, f))
})
names(seurat_list)<- gsub("_seurat.Rds", "", files)


cd4<- seurat_list$cd4
cd4$detailed_annot<- ifelse(cd4$Treatment=="Untreated", cd4$detailed_annot_byNMFrank9, cd4$detailed_annot_projected_manual)

cd8<- seurat_list$cd8
cd8$detailed_annot<- ifelse(cd8$Treatment=="Untreated", cd8$detailed_annot_byNMFrank9, cd8$detailed_annot_projected_manual)

myeloid<- seurat_list$myeloid
myeloid@meta.data<- myeloid@meta.data[,!colnames(myeloid@meta.data) %in% c("raw_clonotype_id", "v_gene", "j_gene", "clonotype_size",
                                                                           "clonal", "clonotype_size_max20", "macrophage1", "microglia1",
                                                                           "broad_annot_projected", "broad_annot_projected_manual")]
myeloid$broad_annot<-"Myeloid"
myeloid$detailed_annot<- "Myeloid"

seurat_list$cd4<-cd4; seurat_list$cd8<-cd8; seurat_list$myeloid<- myeloid

seurat_list<- lapply(seurat_list, function(x){
  x@meta.data<- x@meta.data[,!grepl("NMF", colnames(x@meta.data)) & 
                              !grepl("tmp", colnames(x@meta.data)) & 
                              !grepl("cd4_expr", colnames(x@meta.data)) &
                              !grepl("cd8_expr", colnames(x@meta.data)) &
                              !grepl("detailed_annot_projected_manual", colnames(x@meta.data))]
  return(x)
})

saveRDS(seurat_list$cd4, file=paste0(data_dir, "cd4_seurat.Rds"))
saveRDS(seurat_list$cd8, file=paste0(data_dir, "cd8_seurat.Rds"))
saveRDS(seurat_list$myeloid, file=paste0(data_dir, "myeloid_seurat.Rds"))
```



## DE between programs by treatment
```{r}

de_list<-lapply(seurat_list, function(seurat){
  ## Identify programs to test- >3 cells in program in both treated/untreated
  programs_in_both<- table(seurat$Treatment, seurat$Program_projected)>3
  programs_in_both<- colnames(programs_in_both)[colSums(programs_in_both)==2]
  message(paste0("Programs testing: ", paste(programs_in_both, collapse = ", ")))
  
  ## Cycle through each program shared by treated/untreated (>3 cells in program)
  programs_de<- lapply(programs_in_both, function(program){
    print(program)
    seurat_subset<- subset(seurat, Program_projected==program)
    seurat_subset<- SetIdent(seurat_subset, value=seurat_subset$Treatment)
    markers<- FindMarkers(seurat_subset, ident.1 = "Untreated", ident.2="PD1")
    markers$program<- program
    markers$gene<- rownames(markers)
    return(markers)
  })
  programs_de<- do.call("rbind", programs_de)
  
  return(programs_de)
})

## Remove any cell types with no degs
de_list<- de_list[lapply(de_list, class)=="data.frame"]


saveRDS(de_list, file=paste0(analysis_dir, "DE_Untreated.v.PD1_byProgram.Rds"))
  
```
